pyproject.toml
[project]
name = "api-automation-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "allure-pytest>=2.14.2",
    "assertpy>=1.1",
    "pydantic>=2.11.4",
    "pytest>=8.3.5",
    "requests>=2.32.3",
]
